Move update method to document trait
